requirements.txt
requests
pandas
tqdm
python-dotenv
openpyxl